% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_names.R
\name{get_dataset_names}
\alias{get_dataset_names}
\title{get_dataset_names()}
\usage{
get_dataset_names(filter_str = NULL)
}
\arguments{
\item{filter_str}{A character string by which to filter the dataset names
containing the string.}
}
\value{
A data.table
}
\description{
Function produces a data.table/dataframe of the
Census Bureau's dataset names that can be used in other \code{Rcensus::}
functions calling for a dataset name. See Census Bureau's publicly available
\href{https://www.census.gov/data/developers/data-sets.html}{datasets} for
descriptions.
}
\details{
Get the names and descriptions of the Census Bureau's datasets.
}
\author{
Rick Dean
}
