% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{get_groups()}
\usage{
get_groups(dataset, vintage = NULL)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5")}

\item{vintage}{An optional numeric that sets the year of interest.}
}
\value{
A data.table
}
\description{
Function produces a data.table/data frame of
variable groups and their descriptive parameters.
}
\details{
Get the names of Census Bureau variable groups and their descriptive parameters
}
\author{
Rick Dean
}
