% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_vintage_data.R
\name{get_multi_vintage_data}
\alias{get_multi_vintage_data}
\title{get_multi_vintage_data()}
\usage{
get_multi_vintage_data(
  dataset,
  vintage_v,
  vars,
  vars_init = c("GEO_ID", "NAME"),
  region = NULL,
  regionin = NULL,
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5")}

\item{vintage_v}{A numeric vector that sets the vintages of interest.}

\item{vars}{Can be either a string vector of variable names to be acquired.
Also can be a named list of string vectors to separate groups of variables.}

\item{vars_init}{A string vector of variable names to append to \code{vars}, The default
is \code{c("GEO_ID", "NAME")}. Set it to NULL to disregard it.}

\item{region}{A string that specifies the geography of the request.
See \code{Rcensus::get_geography()} for assistance in obtaining
these values.}

\item{regionin}{A string that sets a qualifier for \code{region}.}

\item{key}{A key string issued by the Census Bureau in making data requests.}
}
\value{
A data.table
}
\description{
Function produces a data.table/dataframe of
Census Bureau data for multiple vintages.  The function requires an access key
issued from the Bureau.
}
\details{
Get Census Bureau data for a specific dataset, variables, and region
in the form of a dataframe for multiple vintages.
}
\author{
Rick Dean
}
