% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vintage_data.R
\name{get_vintage_data}
\alias{get_vintage_data}
\title{get_vintage_data()}
\usage{
get_vintage_data(
  dataset,
  vintage = NULL,
  vars,
  vars_init = c("GEO_ID", "NAME"),
  group = NULL,
  group_values = c("estimate", "moe"),
  region = NULL,
  regionin = NULL,
  na_cols = NULL,
  shape = "long",
  melt_meas = NULL,
  melt_values = NULL,
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5")}

\item{vintage}{An optional numeric that sets the vintage of interest.}

\item{vars}{Can be either a string vector of variable names to be acquired.
Also can be a named list of string vectors to separate groups of variables.}

\item{vars_init}{A string vector of variable names to append to \code{vars}, The default
is \code{c("GEO_ID", "NAME")}. Set it to NULL to disregard it.}

\item{group}{A string that names the group that has a collection of variables. A dataframe
is returned in the "long" format with a factor column named "variables" with their respective
values.}

\item{group_values}{A string vector that sets the type of output values.
Permissible values are "estimate" (for estimates) "moe" (for margin of error).
The default is \code{c("estimate", "moe")} for both.}

\item{region}{A string that specifies the geography of the request.
See \code{Rcensus::get_geography()} for assistance in obtaining
these values.}

\item{regionin}{A string that sets a qualifier for \code{region}.}

\item{na_cols}{If TRUE will remove all rows with missing values. If a
vector of column names/integers will check their values for missing values.}

\item{shape}{A string that sets the shape of the returned dataframe. Permissible
values are "long" (the default) or "wide".}

\item{melt_meas}{A \code{list} that contains character vectors of column names/indexes from
\code{dataset} to be consolidated. It is the \code{measure.vars} argument in data.table's
wide-to-long \href{https://rdatatable.gitlab.io/data.table/reference/melt.data.table.html}{melt}
reshaping tool.}

\item{melt_values}{A Vector that assigns a value to each of the consolidated columns.}

\item{key}{A key string issued by the Census Bureau in making data requests.}
}
\value{
A data.table
}
\description{
Function produces a data.table/dataframe of
Census Bureau data.  The function requires an access key
issued from the Bureau.
}
\details{
Get Census Bureau data for a specific dataset, variables, and region
in the form of a dataframe.
}
\author{
Rick Dean
}
