% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{get_variables()}
\usage{
get_variables(dataset, vintage = NULL, group = NULL, brief = FALSE)
}
\arguments{
\item{dataset}{A string that sets the name of the dataset of interest (e.g. "acs/acs5").
See \code{Rcensus::get_dataset_names()} for available dataset names.}

\item{vintage}{An optional numeric that sets the year of interest.}

\item{group}{An optional string that sets the group name associated with a set of variables}

\item{brief}{If TRUE out of the 10 columns, will return only columns \code{name}, \code{label}, \code{concept}, \code{predicateType}.}
}
\value{
A data.table
}
\description{
Function produces a data.table/data frame of variables and their
descriptive parameters from the Census Bureau's publicly available
\href{https://www.census.gov/data/developers/data-sets.html}{datasets}.
}
\details{
Get the names of Census Bureau variables and their descriptive parameters
}
\author{
Rick Dean
}
