% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vintage_data.R
\name{get_vintage_data}
\alias{get_vintage_data}
\title{get_vintage_data()}
\usage{
get_vintage_data(
  dataset,
  vintage = NULL,
  vars,
  group = NULL,
  region = NULL,
  regionin = NULL,
  na_cols = NULL,
  suffixes = NULL,
  melt_meas = NULL,
  melt_var_name = "variable",
  melt_val_name = "value",
  dcast_for = NULL,
  dcast_var = NULL,
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{dataset}{A string that sets the name of the data set of interest (e.g. "acs/acs5")}

\item{vintage}{An optional numeric that sets the year of interest.}

\item{vars}{A string vector of variable names to be acquired.}

\item{group}{A string that names the group that has a collection of variables. A dataframe
is returned in the "long" format with a factor column named "variables" with their respective
values.}

\item{region}{A string that specifies the geography of the request.
See \code{Rcensus::get_geography()} for assistance in obtaining
these values.}

\item{regionin}{A string that sets a qualifier for \code{region}.}

\item{na_cols}{If TRUE will remove all rows with missing values. If a
vector of column names/integers will check their values for missing values.}

\item{suffixes}{A vector string where column names ending with the suffixes will be kept.}

\item{melt_meas}{A \code{list} that contains character vectors of column names/indexes from
\code{dataset} to be consolidated. It is the \code{measure.vars} argument in data.table's
wide-to-long \href{https://rdatatable.gitlab.io/data.table/reference/melt.data.table.html}{melt}
reshaping tool.}

\item{melt_var_name}{A character string (or character vector) that defines the new name(s) of the
consolidated variable name column.}

\item{melt_val_name}{A character string (or character vector) that defines the new name(s) of the
consolidated value column(s) defined in \code{melt_meas} above. It is the \code{value.name} argument
in data.table's wide-to-long reshaping tool.}

\item{dcast_for}{The LHS ~ RHS formula to use in data.table's
\href{https://rdatatable.gitlab.io/data.table/reference/dcast.data.table.html}{decast}
long-to-wide reshaping tool.}

\item{dcast_var}{A vector of column names/indexes that defines the \code{value.var}
argument in data.table's long-to-wide dcast.}

\item{key}{A key string issued by the Census Bureau in making data requests.}
}
\value{
A data.table
}
\description{
Function produces a data.table/dataframe of
Census Bureau data.  The function requires an access key
issued from the Bureau.
}
\details{
Get Census Bureau data for a specific dataset, variables, and region
in the form of a dataframe.
}
\author{
Rick Dean
}
